+++ lib/processes/blockchain/txs/bond2spend.ex
