+++ lib/processes/blockchain/txs/channel_block.ex
