+++ lib/processes/blockchain/txs/reveal.ex
