+++ lib/processes/blockchain/txs/sign_transaction.ex
