+++ lib/processes/blockchain/txs/slasher.ex
