+++ lib/processes/blockchain/txs/wait2bond.ex
