+++ lib/pure/det_hash.ex
