+++ lib/pure/hash_math.ex
