+++ lib/pure/verify_balances.ex
