+++ lib/pure/verify_tx.ex
